+++ baseapp/src/main/java/com/ljwx/baseapp/page/IPageLocalEvent.kt
@@ -7,7 +7,7 @@ interface IPageLocalEvent : ISendLocalEvent {
 
     fun registerLocalEvent(
         action: String?,
-        observer: (action: String, type: Int?, intent: Intent) -> Unit
+        observer: (action: String, type: Long?, intent: Intent) -> Unit
     )
 
     fun unregisterLocalEvent(action: String?)
