+++ baseapp/src/main/java/com/ljwx/baseapp/page/IPageRefreshLayout.kt
@@ -18,9 +18,9 @@ interface IPageRefreshLayout {
      * 触发刷新
      */
     @Deprecated(message = "deprecated")
-    fun onRefreshData(type: Int = 0)
+    fun onRefreshData(type: Long = 0)
 
-    fun onLoadData(type: Int = 0)
+    fun onLoadData(type: Long = 0)
 
     /**
      * 刷新结束
